Added ingredients and method to display ingredient

+++ src/Ingredient.java
@@ -5,7 +5,7 @@ public class Ingredient {
 
     public Ingredient(){}
 
-    public Ingredient(String name, int amount, String unit) {
+    public Ingredient(String name, double amount, String unit) {
         this.name = name;
         this.amount = amount;
         this.unit = unit;
